move to the correct folder
