+++ src/Controls/samples/Controls.Sample/Pages/Blazor/BlazorPage.xaml.cs
@@ -5,8 +5,8 @@ using Microsoft.Maui.Controls;
 using Microsoft.Maui.Graphics;
 
 #if NET6_0_OR_GREATER
+using Maui.Controls.Sample.Pages.Blazor;
 using Microsoft.AspNetCore.Components.Web;
-using Maui.Controls.Sample.Pages.Others;
 using Microsoft.AspNetCore.Components.WebView.Maui;
 #endif
