+++ src/Core/src/Hosting/Internal/MauiApplicationServiceCollection.cs
@@ -3,7 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Microsoft.Maui
+namespace Microsoft.Maui.Hosting
 {
 	internal sealed class MauiApplicationServiceCollection : IServiceCollection
 	{
